+++ python/pathway/io/_utils.py
@@ -389,17 +389,12 @@ def construct_s3_data_storage(
 def check_raw_and_plaintext_only_kwargs_for_message_queues(f):
     @functools.wraps(f)
     def wrapper(*args, **kwargs):
-        if kwargs.get("format") not in ("raw", "plaintext"):
-            unexpected_params = [
-                "key",
-                "value",
-                "headers",
-            ]
-            for param in unexpected_params:
-                if param in kwargs and kwargs[param] is not None:
-                    raise ValueError(
-                        f"Unsupported argument for {format} format: {param}"
-                    )
+        data_format = kwargs.get("format")
+        if data_format not in ("raw", "plaintext"):
+            if "value" in kwargs and kwargs["value"] is not None:
+                raise ValueError(
+                    f"Unsupported argument for {data_format} format: 'value'"
+                )
 
         return f(*args, **kwargs)
 
@@ -427,31 +422,42 @@ class MessageQueueOutputFormat:
     ) -> MessageQueueOutputFormat:
         key_field_index = None
         header_fields: dict[str, int] = {}
-        if format == "json":
-            data_format = api.DataFormat(
-                format_type="jsonlines",
-                key_field_names=[],
-                value_fields=_format_output_value_fields(table),
+        extracted_field_indices: dict[str, int] = {}
+        columns_to_extract: list[ColumnReference] = []
+        allowed_column_types = (dt.BYTES, dt.STR, dt.ANY)
+
+        # Common part for all formats: obtain key field index and prepare header fields
+        if key is not None:
+            if table[key._name]._column.dtype not in allowed_column_types:
+                raise ValueError(
+                    f"The key column should be of the type '{allowed_column_types[0]}'"
+                )
+            key_field_index = cls.add_column_reference_to_extract(
+                key, columns_to_extract, extracted_field_indices
             )
-        elif format == "dsv":
+        if headers is not None:
+            for header in headers:
+                header_fields[header.name] = cls.add_column_reference_to_extract(
+                    header, columns_to_extract, extracted_field_indices
+                )
+
+        # Format-dependent parts: handle json and dsv separately
+        if format == "json" or format == "dsv":
+            for column_name in table._columns:
+                cls.add_column_reference_to_extract(
+                    table[column_name], columns_to_extract, extracted_field_indices
+                )
+            table = table.select(*columns_to_extract)
             data_format = api.DataFormat(
-                format_type="dsv",
+                format_type="jsonlines" if format == "json" else "dsv",
                 key_field_names=[],
                 value_fields=_format_output_value_fields(table),
                 delimiter=delimiter,
             )
         elif format == "raw" or format == "plaintext":
             value_field_index = None
-            extracted_field_indices: dict[str, int] = {}
-            columns_to_extract: list[ColumnReference] = []
-            allowed_column_types = (dt.BYTES if format == "raw" else dt.STR, dt.ANY)
-
-            if key is not None:
-                if value is None:
-                    raise ValueError("'value' must be specified if 'key' is not None")
-                key_field_index = cls.add_column_reference_to_extract(
-                    key, columns_to_extract, extracted_field_indices
-                )
+            if key is not None and value is None:
+                raise ValueError("'value' must be specified if 'key' is not None")
             if value is not None:
                 value_field_index = cls.add_column_reference_to_extract(
                     value, columns_to_extract, extracted_field_indices
@@ -468,21 +474,7 @@ class MessageQueueOutputFormat:
                     value, columns_to_extract, extracted_field_indices
                 )
 
-            if headers is not None:
-                for header in headers:
-                    header_fields[header.name] = cls.add_column_reference_to_extract(
-                        header, columns_to_extract, extracted_field_indices
-                    )
-
             table = table.select(*columns_to_extract)
-
-            if (
-                key is not None
-                and table[key._name]._column.dtype not in allowed_column_types
-            ):
-                raise ValueError(
-                    f"The key column should be of the type '{allowed_column_types[0]}'"
-                )
             if table[value._name]._column.dtype not in allowed_column_types:
                 raise ValueError(
                     f"The value column should be of the type '{allowed_column_types[0]}'"