+++ tests/integration/test_dsv_output.rs
@@ -29,7 +29,7 @@ fn test_dsv_format_ok() -> eyre::Result<()> {
 
     let target_payloads = vec![b"b;c;time;diff".to_vec(), b"\"x\";\"y\";0;1".to_vec()];
 
-    assert_eq!(result.values.len(), 0);
+    assert_eq!(result.values, &[Value::from("x"), Value::from("y")]);
     assert_eq!(result.payloads.len(), target_payloads.len());
     for (result_payload, target_payload) in zip(result.payloads, target_payloads) {
         assert_document_raw_byte_contents(&result_payload, &target_payload);