+++ tests/integration/test_json_output.rs
@@ -23,7 +23,7 @@ fn test_json_format_ok() -> eyre::Result<()> {
         &result.payloads[0],
         r#"{"a":"b","diff":1,"time":0}"#.as_bytes(),
     );
-    assert_eq!(result.values.len(), 0);
+    assert_eq!(result.values, &[Value::from("b")]);
 
     Ok(())
 }
@@ -43,7 +43,7 @@ fn test_json_int_serialization() -> eyre::Result<()> {
         &result.payloads[0],
         r#"{"a":555,"diff":1,"time":0}"#.as_bytes(),
     );
-    assert_eq!(result.values.len(), 0);
+    assert_eq!(result.values.len(), 1);
 
     Ok(())
 }
@@ -63,7 +63,7 @@ fn test_json_float_serialization() -> eyre::Result<()> {
         &result.payloads[0],
         r#"{"a":5.55,"diff":1,"time":0}"#.as_bytes(),
     );
-    assert_eq!(result.values.len(), 0);
+    assert_eq!(result.values.len(), 1);
 
     Ok(())
 }
@@ -83,7 +83,7 @@ fn test_json_bool_serialization() -> eyre::Result<()> {
         &result.payloads[0],
         r#"{"a":true,"diff":1,"time":0}"#.as_bytes(),
     );
-    assert_eq!(result.values.len(), 0);
+    assert_eq!(result.values.len(), 1);
 
     Ok(())
 }
@@ -103,7 +103,7 @@ fn test_json_null_serialization() -> eyre::Result<()> {
         &result.payloads[0],
         r#"{"a":null,"diff":1,"time":0}"#.as_bytes(),
     );
-    assert_eq!(result.values.len(), 0);
+    assert_eq!(result.values.len(), 1);
 
     Ok(())
 }
@@ -124,7 +124,7 @@ fn test_json_pointer_serialization() -> eyre::Result<()> {
         &result.payloads[0],
         r#"{"a":"^04000000000000000000000000","diff":1,"time":0}"#.as_bytes(),
     );
-    assert_eq!(result.values.len(), 0);
+    assert_eq!(result.values.len(), 1);
 
     Ok(())
 }
@@ -149,7 +149,7 @@ fn test_json_tuple_serialization() -> eyre::Result<()> {
         &result.payloads[0],
         r#"{"a":[true,null],"diff":1,"time":0}"#.as_bytes(),
     );
-    assert_eq!(result.values.len(), 0);
+    assert_eq!(result.values.len(), 1);
 
     Ok(())
 }
@@ -171,7 +171,7 @@ fn test_json_date_time_naive_serialization() -> eyre::Result<()> {
         &result.payloads[0],
         r#"{"a":"2023-05-15T10:51:00.000000000","diff":1,"time":0}"#.as_bytes(),
     );
-    assert_eq!(result.values.len(), 0);
+    assert_eq!(result.values.len(), 1);
 
     Ok(())
 }
@@ -191,7 +191,7 @@ fn test_json_date_time_utc_serialization() -> eyre::Result<()> {
         &result.payloads[0],
         r#"{"a":"2023-05-15T10:51:00.000000000+0000","diff":1,"time":0}"#.as_bytes(),
     );
-    assert_eq!(result.values.len(), 0);
+    assert_eq!(result.values.len(), 1);
 
     Ok(())
 }
@@ -212,7 +212,7 @@ fn test_json_duration_serialization() -> eyre::Result<()> {
         &result.payloads[0],
         r#"{"a":1197780000000000,"diff":1,"time":0}"#.as_bytes(),
     );
-    assert_eq!(result.values.len(), 0);
+    assert_eq!(result.values.len(), 1);
 
     Ok(())
 }